+++ NEWS.md
@@ -1,3 +1,8 @@
+- Version 0.11.0 - released 2020-08-27
+  - `--clipboard` option removed in favour of `--clipboard-cmd`
+  - `--dmenu-prompt` option and trailing dmenu arguments removed in favour of
+    `--menu-cmd`
+
 - Version 0.10.0 - released 2020-08-02
   - Added support for arbitrary clipboard commands - thanks @ign0tus!
   - `--clipboard` option deprecated in favour of `--clipboard-cmd`